Began implementing mobile app tracking functionality.
The app has the ability to parse XML from the server and can use it to create a marker. However, it is currently creating only the first marker. The app also needs to be tested with moving vehicles to see if it can properly update the markers with movement, but it probably can't.