Added route and stop selection to the android app.
The key activity now displays two buttons that can be used to access
lists of routes and stops. Selecting a route in the list will add it to
the map if it was not being displayed, and selecting it again will
remove it. Multiple routes can be displayed at a time. Selecting a stop
in this way turns it blue; originally, I was going to have the markers
bounce to match the website, but marker bouncing is much, much more
involved in Android than in the web version. Markers have also all been
updated to match the website, with a blue stop sign used for selected
stops.
While this functionality is completed, the UI is not 100% complete and
I'll fix that when I'm more awake.

Another, smaller change to the app is that the activities now allow only
one instance of themselves to exist at a time. Instead of having
instances stack up (Map, Key, Map, Help, Key, About, Key, etc.),
creating a new instance destroys the existing instance. This makes
navigating the app with the back button easier, and probably makes it
more efficient somehow.